Save to file and "html.parser"
Thanks for the code!
I was searching for a python lyric fetcher just like this one. I added a few methods to suit my task. I think you should see them and, if you want it, add to your code, since someone else may need it. I added the option to save lyrics, tracks and albums in a file, for other projects to use them. I didn't solve the problem, but for me the method getTracks fails, because does not find 'album' in html, so i couldn't test the method saveTracks, so i recommend that before you add, test it (if getTracks works for you).
I also added the argument "html.parser" on every call of BeautifulSoup, since in Python 3.5 (the on i used) it raises a little warning.
Cheers,
Diogo Ferreira